+++ example/calc/calc/src/main/java/mb/calc/util/AnalyzedStrategoTransformTaskDef.java
@@ -22,6 +22,6 @@ public abstract class AnalyzedStrategoTransformTaskDef extends StrategoTransform
     }
 
     @Override protected IStrategoTerm getAst(ExecContext context, ConstraintAnalyzeTaskDef.Output input) {
-        return input.result.ast;
+        return input.result.analyzedAst;
     }
 }