+++ example/tiger/manual/tiger/src/test/java/mb/tiger/TigerConstraintAnalyzerTest.java
@@ -22,7 +22,7 @@ class TigerConstraintAnalyzerTest extends TestBase {
         final ReadableResource file = textFile("a.tig", "1 + nil");
         final JsglrParseOutput parsed = parse(file);
         final SingleFileResult result = analyze(file, parsed.ast);
-        assertNotNull(result.ast);
+        assertNotNull(result.analyzedAst);
         assertNotNull(result.analysis);
         assertTrue(result.messages.containsError());
     }
@@ -31,7 +31,7 @@ class TigerConstraintAnalyzerTest extends TestBase {
         final ReadableResource file = textFile("a.tig", "1 + 2");
         final JsglrParseOutput parsed = parse(file);
         final SingleFileResult result = analyze(file, parsed.ast);
-        assertNotNull(result.ast);
+        assertNotNull(result.analyzedAst);
         assertNotNull(result.analysis);
         assertTrue(result.messages.isEmpty());
     }
@@ -51,15 +51,15 @@ class TigerConstraintAnalyzerTest extends TestBase {
         final MultiFileResult result = analyze(MapView.of(asts));
         final ConstraintAnalyzer.@Nullable Result result1 = result.getResult(file1.getKey());
         assertNotNull(result1);
-        assertNotNull(result1.ast);
+        assertNotNull(result1.analyzedAst);
         assertNotNull(result1.analysis);
         final ConstraintAnalyzer.@Nullable Result result2 = result.getResult(file2.getKey());
         assertNotNull(result2);
-        assertNotNull(result2.ast);
+        assertNotNull(result2.analyzedAst);
         assertNotNull(result2.analysis);
         final ConstraintAnalyzer.@Nullable Result result3 = result.getResult(file3.getKey());
         assertNotNull(result3);
-        assertNotNull(result3.ast);
+        assertNotNull(result3.analyzedAst);
         assertNotNull(result3.analysis);
         assertEquals(1, result.messages.size());
         assertTrue(result.messages.containsError());
@@ -84,15 +84,15 @@ class TigerConstraintAnalyzerTest extends TestBase {
         final MultiFileResult result = analyze(MapView.of(asts));
         final ConstraintAnalyzer.@Nullable Result result1 = result.getResult(file1.getKey());
         assertNotNull(result1);
-        assertNotNull(result1.ast);
+        assertNotNull(result1.analyzedAst);
         assertNotNull(result1.analysis);
         final ConstraintAnalyzer.@Nullable Result result2 = result.getResult(file2.getKey());
         assertNotNull(result2);
-        assertNotNull(result2.ast);
+        assertNotNull(result2.analyzedAst);
         assertNotNull(result2.analysis);
         final ConstraintAnalyzer.@Nullable Result result3 = result.getResult(file3.getKey());
         assertNotNull(result3);
-        assertNotNull(result3.ast);
+        assertNotNull(result3.analyzedAst);
         assertNotNull(result3.analysis);
         assertTrue(result.messages.isEmpty());
     }