+++ lwb/spoofax.lwb.eclipse/src/main/java/mb/spoofax/lwb/eclipse/dynamicloading/DynamicEditorTracker.java
@@ -74,6 +74,13 @@ public class DynamicEditorTracker extends WindowAndPartListener {
         });
     }
 
+    @Override public void unregister() {
+        deactivateContext();
+        super.unregister();
+        this.editors.clear();
+        this.contextService = null;
+    }
+
 
     // IWindowListener implementation